requirements.txt
alembic==1.14.0
colorlog==6.9.0
imbalanced-learn==0.12.4
joblib==1.4.2
Mako==1.3.6
MarkupSafe==3.0.2
numpy==2.1.3
optuna==4.1.0
packaging==24.2
pandas==2.2.3
pandas_ta==0.3.14b0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
SQLAlchemy==2.0.36
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
xgboost==2.1.3
